+++ R/ClusterR.R
@@ -1,6 +1,5 @@
 #' @useDynLib ClusterR, .registration = TRUE
 #' @importFrom Rcpp evalCpp
-#' @import gtools
 #' @importFrom utils globalVariables
 #' @importFrom lifecycle badge deprecate_warn is_present
 NULL